requirements.txt
# Dependencias para el proyecto Instituto_virtual
# Flask: framework web usado en app.py
# mysql-connector-python: conector MySQL (importado como mysql.connector en database.py)
Flask>=2.2
psycopg2-binary>=2.9
gunicorn
python-dotenv>=1.0